Write Golang comments; Download enhancements
Add a function to write a string formatted as a Go comment (including
indentation and word-wrapping).

Use the Last-Modified header to determine if we should skip
re-downloading a file, rather than depend on the server to set the
correct HTTP status based on If-Modified-Since in the request.

Download() will now try to use the filename from the
Content-Disposition header, if present.

Add an optional argument to Download() to override the output
filename, rather than take it from the URL or Content-Disposition.