Fix some issues with GolangCommentWriter
The Write() method in the previous commit slices over the input []byte
for efficiency, but this doesn't work if there's unflushed data upon
return.  We need to revert to making a copy of the input data when a
single word spans multiple calls to Write().

We also need a mutex, since neither Write() nor Close() are
thread-safe.